backend/requirements.txt
flask
python-dotenv
requests
flask-cors
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-Login